+++ src/test/java/org/delia/zdb/ZDBTests.java
@@ -1,0 +1,21 @@
+package org.delia.zdb;
+
+import static org.junit.Assert.assertEquals;
+
+import org.junit.Before;
+import org.junit.Test;
+
+public class ZDBTests  {
+	
+	@Test
+	public void testTool() {
+		assertEquals(1,2);
+	}
+	
+	// --
+
+	@Before
+	public void init() {
+	}
+	
+}
