+++ src/test/java/org/delia/zdb/core/ZDBIContext.java
@@ -1,0 +1,7 @@
+package org.delia.zdb.core;
+
+//holds session data regarding db
+public class ZDBIContext {
+	public Object connObject; //for internal use only
+	public boolean disableSqlLogging; //for internal use only
+}
