+++ src/test/java/org/delia/zdb/core/ZDBInterfaceFactory.java
@@ -1,0 +1,14 @@
+package org.delia.zdb.core;
+
+import org.delia.db.DBCapabilties;
+import org.delia.db.DBType;
+
+public interface ZDBInterfaceFactory {
+	DBType getDBType();
+	DBCapabilties getCapabilities();
+	
+	ZDBConnection openConnection();
+	
+	boolean isSQLLoggingEnabled();
+	void enableSQLLogging(boolean b);
+}
